PWN-5316 - changes according to comments
