+++ app/src/main/java/org/p2p/wallet/solend/ui/earn/SolendEarnFragment.kt
@@ -15,7 +15,7 @@ import org.koin.android.ext.android.inject
 import org.p2p.uikit.utils.attachAdapter
 import org.p2p.wallet.R
 import org.p2p.wallet.common.mvp.BaseMvpFragment
-import org.p2p.wallet.common.ui.widget.EarnWidgetState
+import org.p2p.wallet.common.ui.widget.earnwidget.EarnWidgetState
 import org.p2p.wallet.databinding.FragmentSolendEarnBinding
 import org.p2p.wallet.solend.model.SolendDepositToken
 import org.p2p.wallet.solend.ui.earn.adapter.SolendEarnAdapter
@@ -48,53 +48,7 @@ class SolendEarnFragment :
             tokensRecyclerView.attachAdapter(earnAdapter)
 
             // Update after real integration
-            viewEarnWidget.setWidgetState(EarnWidgetState.LearnMore)
-            viewEarnWidget.setOnButtonClickListener {
-                timerJob?.cancel()
-                when (val state = viewEarnWidget.currentState) {
-                    is EarnWidgetState.Balance -> {
-                        viewEarnWidget.setWidgetState(
-                            EarnWidgetState.Depositing(R.string.earn_widget_deposit_withdrawing)
-                        )
-                    }
-                    is EarnWidgetState.Depositing -> {
-                        if (state.buttonTextRes == R.string.earn_widget_deposit_withdrawing) {
-                            viewEarnWidget.setWidgetState(EarnWidgetState.DepositFoundsFailed)
-                        } else {
-                            val timer = (1..Int.MAX_VALUE)
-                                .asSequence()
-                                .asFlow()
-                                .onEach { delay(1_000) }
-
-                            val job = Job()
-                            val uiScope = CoroutineScope(Dispatchers.Main + job)
-                            timerJob = uiScope.launch() {
-                                var balance = 89.5708912487
-                                timer.collect {
-                                    balance += (it.toFloat() / 100000000000) // + Random.nextInt(1, 9)
-                                    viewEarnWidget.setWidgetState(EarnWidgetState.Balance(BigDecimal(balance)))
-                                }
-                            }
-                        }
-                    }
-                    EarnWidgetState.LearnMore -> {
-                        viewEarnWidget.setWidgetState(
-                            EarnWidgetState.Depositing(R.string.earn_widget_deposit_sending)
-                        )
-                    }
-                    EarnWidgetState.DepositFoundsFailed -> {
-                        viewEarnWidget.setWidgetState(
-                            EarnWidgetState.Error(
-                                R.string.earn_widget_error_message_withdrawal,
-                                R.string.earn_widget_error_button_ok
-                            )
-                        )
-                    }
-                    is EarnWidgetState.Error -> {
-                        viewEarnWidget.setWidgetState(EarnWidgetState.LearnMore)
-                    }
-                }
-            }
+            setupMockWidgetData()
         }
 
         presenter.load()
@@ -114,5 +68,55 @@ class SolendEarnFragment :
     override fun onDestroyView() {
         timerJob?.cancel()
         super.onDestroyView()
+    }
+
+    private fun FragmentSolendEarnBinding.setupMockWidgetData() {
+        viewEarnWidget.setWidgetState(EarnWidgetState.LearnMore)
+        viewEarnWidget.setOnButtonClickListener {
+            timerJob?.cancel()
+            when (val state = viewEarnWidget.currentState) {
+                is EarnWidgetState.Balance -> {
+                    viewEarnWidget.setWidgetState(
+                        EarnWidgetState.Depositing(R.string.earn_widget_deposit_withdrawing)
+                    )
+                }
+                is EarnWidgetState.Depositing -> {
+                    if (state.buttonTextRes == R.string.earn_widget_deposit_withdrawing) {
+                        viewEarnWidget.setWidgetState(EarnWidgetState.DepositFoundsFailed)
+                    } else {
+                        val timer = (1..Int.MAX_VALUE)
+                            .asSequence()
+                            .asFlow()
+                            .onEach { delay(1_000) }
+
+                        val job = Job()
+                        val uiScope = CoroutineScope(Dispatchers.Main + job)
+                        timerJob = uiScope.launch() {
+                            var balance = 89.5708912487
+                            timer.collect {
+                                balance += (it.toFloat() / 100000000000) // + Random.nextInt(1, 9)
+                                viewEarnWidget.setWidgetState(EarnWidgetState.Balance(BigDecimal(balance)))
+                            }
+                        }
+                    }
+                }
+                EarnWidgetState.LearnMore -> {
+                    viewEarnWidget.setWidgetState(
+                        EarnWidgetState.Depositing(R.string.earn_widget_deposit_sending)
+                    )
+                }
+                EarnWidgetState.DepositFoundsFailed -> {
+                    viewEarnWidget.setWidgetState(
+                        EarnWidgetState.Error(
+                            R.string.earn_widget_error_message_withdrawal,
+                            R.string.earn_widget_error_button_ok
+                        )
+                    )
+                }
+                is EarnWidgetState.Error -> {
+                    viewEarnWidget.setWidgetState(EarnWidgetState.LearnMore)
+                }
+            }
+        }
     }
 }
