+++ app/src/main/res/layout/widget_earn_view.xml
@@ -66,9 +66,9 @@
             android:layout_marginHorizontal="20dp"
             android:layout_marginBottom="20dp"
             android:backgroundTint="@color/bg_snow"
-            android:textColor="@color/night"
+            android:textColor="@color/text_night"
             app:iconGravity="end"
-            app:iconTint="@color/night"
+            app:iconTint="@color/icons_night"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
